EBU6307_ZHEYUN_ZHAO_221170559/requirements.txt
pyqt5
numpy
opencv-python
opencv-contrib-python
matplotlib
scipy
scikit-learn
pillow
torch
torchvision